Moved `getReverseHighlight` function to utils
