+++ src/helpers/reverseHighlight.ts
@@ -1,5 +1,5 @@
 import { Hit } from '../types';
-import { getPropertyByPath, unescape } from '../lib/utils';
+import { getPropertyByPath, getReversedHighlight } from '../lib/utils';
 import { TAG_REPLACEMENT } from '../lib/escape-highlight';
 import { component } from '../lib/suit';
 
@@ -13,36 +13,6 @@ export type ReverseHighlightOptions = {
 };
 
 const suit = component('ReverseHighlight');
-
-const getReversedHighlight = (attribute: string) => {
-  const regexHasMark = /(<mark>.*?<\/mark>)/gi;
-  const parts = unescape(attribute).split(regexHasMark);
-
-  if (parts.length === 1) return attribute;
-
-  return parts
-    .reduce((acc: string[], curr) => {
-      if (!curr.length) {
-        acc.push(curr);
-      } else if (regexHasMark.test(curr)) {
-        acc.push(curr.replace(/(<\/mark>|<mark>)/g, ''));
-      } else {
-        const reversed = curr
-          .split(/([^a-z0-9])/gi)
-          .map(escaped =>
-            /(^[a-z0-9]+$)/gi.test(escaped)
-              ? `<mark>${escaped}</mark>`
-              : escaped
-          )
-          .join('');
-
-        acc.push(reversed);
-      }
-
-      return acc;
-    }, [])
-    .join('');
-};
 
 export default function reverseHighlight({
   attribute,
