+++ src/lib/utils/index.ts
@@ -21,6 +21,7 @@ export { default as range } from './range';
 export { default as isEqual } from './isEqual';
 export { default as escape } from './escape';
 export { default as unescape } from './unescape';
+export { default as getReversedHighlight } from './getReversedHighlight';
 export { default as find } from './find';
 export { default as findIndex } from './findIndex';
 export { default as mergeSearchParameters } from './mergeSearchParameters';
